windowing/gbm: split gbm from drm
